fixed so it saves the right information

+++ src/dk/aau/cs/giraf/train/GameLinearLayout.java
@@ -166,7 +166,7 @@ public class GameLinearLayout extends LinearLayout {
             
             int gameID = Integer.parseInt(game[0]);
             int guardianID = Integer.parseInt(game[1]);
-            long childID = Long.valueOf(game[2]);
+            int childID = Integer.valueOf(game[2]);
             String gameName = game[3];
             int TempdistanceBetweenStations = Integer.parseInt(game[4]);
             ArrayList<StationConfiguration> stations = new ArrayList<StationConfiguration>();
